Add files via upload
this file created in plugins/zatca-setting/zatca-settings-form.php and zatca-settings.php. i have created this form in admin panel in woocommerece /setting/zatca-setting when i click copy button , the copy the content of address billing   in this form then we can submit the form and its stored in the db.